refactor: created wrapper function in isUser and removed unnessary backend code
